Remove hamcrest-integration
This target is not used and the module does not exist in
Hamcrest 2.0. In Hamcrest 2.0 the file it contains has been
merged into the hamcrest-core module, which is represented in
Android by the hamcrest/hamcrest-host/hamcrest-hostdex build
targets.

Similarly, in Hamcrest 2.0 the TypeSafeMatcher class is in the
hamcrest-core module and not the hamcrest-library module.

This moves the two classes into the appropriate module in
preparation for upgrading JUnit to 4.12 and Hamcrest to 2.0.

Bug: 33613916
Bug: 30946317
Test: make checkbuild
Change-Id: Ib22cf3dce1180b3df12764093b727ff2482c51e5