Temporarily add void is(Class) to cause compile time errors
The is(Class) method was removed upstream between Hamcrest 1.1
and 1.3 in favour of the instanceOf(Class) method.
Unfortunately, existing usages of it could still compile against
the is(Object) method instead. Although not every existing usage
would compile successfully it is possible that some could and
that would result in a change in the runtime behavior that could
be difficult to detect and fix. This change aims to turn any
significant usage of this method into a compile time error.

Bug: 30946317
Test: make checkbuild
Change-Id: Ib0fac6e944b94ba8f1b5434d64549cc12f20d938